requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
django-environ==0.11.2
anthropic==0.18.1
stripe==7.11.0
django-redis==5.4.0
paypalrestsdk==1.13.1
django-paypal==2.0
